Added movie review statistics.

+++ backend/src/main/java/rs/ac/bg/fon/cinefon/domain/Review.java
@@ -32,5 +32,5 @@ public class Review {
     @JoinColumn(name = "movie_id")
     private Movie movie;
 
-    private byte rating;
+    private int rating;
 }
